Support Certs in the webview, better error handling
